[#1272] Moved submission export tests to own file
